+++ contracts/mockups/redeploys/RealitioMockupV2.sol
@@ -4,15 +4,12 @@ pragma solidity ^0.7.5;
 import "hardhat/console.sol";
 
 // a mockup to test changing the proxy, this is as per the original but always says the winner is 69
-contract RealitioMockupV2
-
-{
-    uint result = 420;
+contract RealitioMockupV2 {
+    uint256 result = 420;
     bytes32 public actualQuestionId;
     bytes32 public actualContentHash;
 
-    function setResult(uint _result) public
-    {
+    function setResult(uint256 _result) public {
         if (_result == 69) {
             result = (type(uint256).max);
         } else {
@@ -20,7 +17,14 @@ contract RealitioMockupV2
         }
     }
 
-    function askQuestion(uint256 template_id, string calldata question, address arbitrator, uint32 timeout, uint32 opening_ts, uint256 nonce) external payable returns (bytes32) {
+    function askQuestion(
+        uint256 template_id,
+        string calldata question,
+        address arbitrator,
+        uint32 timeout,
+        uint32 opening_ts,
+        uint256 nonce
+    ) external payable returns (bytes32) {
         bytes32 content_hash = keccak256(abi.encodePacked(template_id, opening_ts, question));
         bytes32 question_id = keccak256(abi.encodePacked(content_hash, arbitrator, timeout, msg.sender, nonce));
         actualContentHash = content_hash;
@@ -33,7 +37,7 @@ contract RealitioMockupV2
     function resultFor(bytes32 question_id) external view returns (bytes32) {
         require(result != 420);
         require(question_id == actualQuestionId);
-        return bytes32(uint(69));
+        return bytes32(uint256(69));
     }
 
     function isFinalized(bytes32 question_id) external view returns (bool) {
@@ -54,6 +58,4 @@ contract RealitioMockupV2
             return 0;
         }
     }
-
 }
-
